added new class SelectFromExamples.cs (extracted tests from ApiExamples.cs)
