+++ src/Coypu.AcceptanceTests/Coypu.AcceptanceTests.csproj
@@ -80,6 +80,7 @@
     <Compile Include="Examples\FindExamples.cs" />
     <Compile Include="Examples\HasExamples.cs" />
     <Compile Include="Examples\ModalDialogExamples.cs" />
+    <Compile Include="Examples\SelectFromExamples.cs" />
     <Compile Include="Examples\ShowsExamples.cs" />
     <Compile Include="Examples\WindowExamples.cs" />
     <Compile Include="Examples\WithinExamples.cs" />
